Polished the code for fetching directory entries.
Now the code for fetching directory entries does indeed fetch directory entries. I removed
minor bugs and modified the procedure for reading the data provided by dir_readdir. The
similar code in unionfs does not work at all. Moreover, it uses some variables names
which are said to be obsolete in bits/dirents.h , that is why I indulged myself with the
liberty to modify the procedure of working with dir_readdir.