Added the code for looking up files.
Extended the set of functions with functions aiding the operation of lookup. At this
moment the code only compiles, however, it implements the fundamental idea of providing
the required information and maintaining a separate, libnetfs independent, tree of light
nodes. This architectures allows fast access to information and rather efficient lazy
updates.